% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{aboutNote}
\alias{aboutNote}
\title{Function to return the note section in the Bulletin Board package}
\usage{
aboutNote()
}
\description{
This function is used to gather all the comments for the About section in one place
}
