% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the dashboard for Vietnam Covid-19}
\usage{
launch_app()
}
\description{
this function will launch the shiny app dashboard that was embedded in the package.
After you installed the package, simply load the package and type the command launch_app() in the console.
}
