% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_function.R
\name{output_ref}
\alias{output_ref}
\title{Create function to generate reference}
\usage{
output_ref(type)
}
\arguments{
\item{type}{The type of references, either 'packages' or 'link'}
}
\description{
This function provides the list of R packages and external resources added in the Reference section
}
\examples{
library(shiny)
output_ref("link")
output_ref("packages")
}
