% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\docType{data}
\name{patient_link}
\alias{patient_link}
\title{Read the csv file for the patient infection route in Vietnam, last updated by 10 Oct 2020}
\format{
An object of class \code{data.frame} with 975 rows and 17 columns.
}
\usage{
patient_link
}
\description{
Return the dataset for patient infection route.
}
\keyword{datasets}
