% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_function.R
\docType{data}
\name{ref}
\alias{ref}
\title{Create function to generate reference}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
ref
}
\arguments{
\item{type}{The type of references, either 'packages' or 'link'}
}
\description{
This function provides the list of R packages used and external resources to add in the Reference section
}
\examples{
library(shiny)
output_ref("link")
output_ref("packages")
}
\keyword{datasets}
