% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiElements.R
\name{aboutNote}
\alias{aboutNote}
\title{Return the note section in the Bulletin Board package}
\usage{
aboutNote()
}
\description{
This function is used to refactor the UI by gathering all the comments for the About section in one place then later deploy to the app.
This function is created for the app's usage only and cannot be called alone in the console.
}
