% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\docType{data}
\name{patient_node}
\alias{patient_node}
\title{Read the csv file for details of infected patients in Vietnam}
\format{
An object of class \code{data.frame} with 1099 rows and 21 columns.
}
\source{
Vietnam's MOH at https://ncov.moh.gov.vn/web/guest/trang-chu
}
\usage{
patient_node
}
\description{
Return the dataset for patient details.
}
\details{
The dataset would include details of each confirmed cases in Vietnam. Main information provided would include the patient id number,
age, address, treatment location, status, infected type, nationality, date announced and death day announced. Other variables would
include the details of vessels that transported oversea patients back home and symtoms if available.
}
\keyword{datasets}
