% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_function.R
\docType{data}
\name{ref}
\alias{ref}
\title{List of references used in the app}
\format{
An object of class \code{data.frame} with 12 rows and 2 columns.
}
\usage{
ref
}
\description{
the ref function returns the list of R packages and external resources refered in the app. As both packages and external links are gathered into one data frame, duplication of values may occur.
}
\examples{
library(tibble)
ref
}
\keyword{datasets}
