% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\docType{data}
\name{vietnam_daily}
\alias{vietnam_daily}
\title{Read the csv file for Vietnam cases by province}
\format{
An object of class \code{data.frame} with 43 rows and 5 columns.
}
\usage{
vietnam_daily
}
\description{
Return the dataset for Vietnam cases and deaths by province updated the latest by 10 Oct 2020
}
\keyword{datasets}
